[Refactor] move all of Types about Masonry in MasonryType.ts file #89
