+++ app/src/lib/common/types/ComponentTypes/ImageType.ts
@@ -1,0 +1,13 @@
+export interface ImagePropsType {
+  src: string;
+  subhead?: string;
+  head?: string;
+  redirectURL?: string;
+  noShowHead?: boolean;
+  zoomWhenHover?: boolean;
+}
+
+export interface ImageItemsStyledPropsType {
+  noShowHead?: boolean;
+  zoomWhenHover?: boolean;
+}
