+++ app/src/lib/common/types/ComponentTypes/MasonryType.ts
@@ -1,0 +1,9 @@
+export interface MasonryStyledPropsType {
+  padding?: string;
+  column?: number;
+}
+
+export interface MasonryPropsType extends MasonryStyledPropsType {
+  id?: string;
+  children: React.ReactNode;
+}
