+++ app/src/lib/components/Masonry/Masonry.tsx
@@ -1,14 +1,8 @@
-import React, { Children } from 'react';
+import { Children } from 'react';
 import styled from 'styled-components';
+import { MasonryPropsType, MasonryStyledPropsType } from '../../common/types/ComponentTypes/MasonryType';
 
-interface Props {
-  id?: string;
-  children: React.ReactNode;
-  padding?: string;
-  column?: number;
-}
-
-const Masonry = ({ id, children, padding, column }: Props) => {
+const Masonry = ({ id, children, padding, column }: MasonryPropsType) => {
   const count = Children.count(children);
 
   if (count <= 6) column = 3;
@@ -24,10 +18,7 @@ const Masonry = ({ id, children, padding, column }: Props) => {
 
 export default Masonry;
 
-const Wrap = styled.div<{
-  padding?: string;
-  column?: number;
-}>`
+const Wrap = styled.div<MasonryStyledPropsType>`
   padding: ${({ padding }) => padding ?? '2em 4em'};
   column-count: ${({ column }) => column ?? 4};
   column-gap: 1.5em;
